python/requirements.txt
pypdf2>=3.0.0
pdf2docx>=0.5.6
reportlab>=4.0.0
python-docx>=1.0.0
openpyxl>=3.1.0
python-pptx>=0.6.21
pytesseract>=0.3.10
Pillow>=10.0.0
baidu-aip>=4.16.0
